aarch64: Add -mveclibabi=sleefgnu
This adds support for the -mveclibabi option to the AArch64 backend of GCC by
implementing the builtin_vectorized_function target hook for AArch64.
The SLEEF Vectorized Math Library's GNUABI interface is used, and
NEON/Advanced SIMD as well as SVE are supported.

This was tested on the gcc testsuite and the llvm-test-suite on a AArch64
host for NEON and SVE as well as on hand-written benchmarks. Where the
vectorization of builtins was applied successfully in loops bound by the
calls to those, significant (>2) performance gains can be observed.

gcc/ChangeLog:

        * config/aarch64/aarch64.opt: Add -mveclibabi option.
        * config/aarch64/aarch64-opts.h: Add aarch64_veclibabi enum.
        * config/aarch64/aarch64-protos.h: Add
        aarch64_builtin_vectorized_function declaration.
        * config/aarch64/aarch64.cc: Handle -mveclibabi option and pure
        scalable type info for scalable vectors without "SVE type" attributes.
        * config/aarch64/aarch64-builtins.cc: Add
        aarch64_builtin_vectorized_function definition.
        * doc/invoke.texi: Document -mveclibabi for AArch64 targets.

gcc/testsuite/ChangeLog:

        * gcc.target/aarch64/vect-vecabi-sleefgnu-neon.c: New testcase.
        * gcc.target/aarch64/vect-vecabi-sleefgnu-sve.c: New testcase.