libstdc++: Convert _RangeAdaptorClosure into a CRTP class [PR108827]
Using the CRTP idiom for this base class avoids bloating the size of a
pipeline when adding distinct empty range adaptor closure objects to it,
as detailed in section 4.1 of P2387R3.

But it means we can no longer define its operator| overloads as hidden
friends, since each instantiation of _RangeAdaptorClosure would then
introduce its own logically different hidden friends.  So for example
during overload resolution for the outer pipe operator in

 :x | (views::reverse | views::join)

we'd have to consider 6 different hidden operator| friends:

  2 from _RangeAdaptorClosure<_Reverse>
  2 from _RangeAdaptorClosure<_Join>
  2 from _RangeAdaptorClosure<_Pipe<_Reverse, _Join>>

which is wasteful and can even cause hard errors in some cases.  So we
instead define the operator| overloads at namespace scope in an isolated
namespace.

	PR libstdc++/108827

libstdc++-v3/ChangeLog:

	* include/std/ranges (__adaptor::_RangeAdaptorClosure): Move ...
	(__adaptor::__closure::_RangeAdaptorClosure): ... here and turn
	it into a CRTP class template.  Move hidden operator| friends
	into namespace scope and adjust their constraints.  Add a
	using-declaration for this at __adaptor::_RangeAdaptorClosure.
	(__closure::__is_range_adaptor_closure_fn): Define.
	(__closure::__is_range_adaptor_closure): Define.
	(__adaptor::_Partial): Adjust use of _RangeAdaptorClosure.
	(__adaptor::_Pipe): Likewise.
	(views::_All): Likewise.
	(views::_Join): Likewise.
	(views::_Common): Likewise.
	(views::_Reverse): Likewise.
	(views::_Elements): Likewise.
	(views::_Adjacent): Likewise.
	(views::_AsRvalue): Likewise.
	(views::_Enumerate): Likewise.
	(views::_AsConst): Likewise.
	* testsuite/std/ranges/adaptors/all.cc: Reintroduce
	static_assert expecting that adding empty range adaptor
	closure objects to a pipeline doesn't increase the size of a
	pipeline.