Return true from operator== for two identical ranges containing NAN.
[Andrew, we talked about this a few months ago.  Just making sure we're
on the same page so I can push it.  Also, a heads-up for Jakub.]

The == operator for ranges signifies that two ranges contain the same
thing, not that they are ultimately equal.  So [2,4] == [2,4], even
though one may be a 2 and the other may be a 3.  Similarly with two
VARYING ranges.

There is an oversight in frange::operator== where we are returning
false for two identical NANs.  This is causing us to never cache NANs
in sbr_sparse_bitmap::set_bb_range.