rust: Disable --enable-languages=rust and silently exclude it from --enable-languages=all for GCC 13
Hi!

As discussed on IRC, gccrs can't compile the standard libraries yet
and requires annoying -frust-incomplete-and-experimental-compiler-do-not-use
option to compile anything.

As such it is better to disable this at least for GCC 13.1, we'd keep
it allowed on the trunk where hopefully support to compile at least
some standard libraries will land soon and eventually a borrow checker.

If enough Rust support is backported from trunk to 13 branch before
13.2, we could revert this change then.

Tested on x86_64-linux with --enable-languages=c,c++,rust and
--enable-languages=c,c++,all , ok for 13 branch?

2023-04-18  Jakub Jelinek  <jakub@redhat.com>

	* configure.ac: For --enable-languages= mentioning explicitly
	rust emit an error, if included in all etc., silently disable
	rust.
	* configure: Regenerated.

	Jakub