Implement range-op entry for sin/cos.
[I don't know why I keep poking at floats.  I must really like the pain.
Jakub, are you OK with this patch for trunk?]

This is the range-op entry for sin/cos.  It is meant to serve as an
example of what we can do for glibc math functions.  It is by no means
exhaustive, just a stub to restrict the return range from sin/cos to
[-1.0, 1.0] with appropriate smarts of NANs.

As can be seen in the testcase, we see sin() as well as
__builtin_sin() in the IL, and can resolve the resulting range
accordingly.

gcc/ChangeLog:

	* gimple-range-op.cc (class cfn_sincos): New.
	(gimple_range_op_handler::maybe_builtin_call): Add case for sin/cos.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/range-sincos.c: New test.