RISC-V: avoid splitting small constants in bcrli_nottwobits patterns
Hi all.

I have noticed that in the case when we try to clear two bits through a
small constant,
and ZBS is enabled then GCC split it into two "andi" instructions.
For example for the following C code:
  int foo(int a) {
    return a & ~ 0x101;
  }

GCC generates the following:
  foo:
     andi a0,a0,-2
     andi a0,a0,-257
     ret

but should be this one:
  foo:
     andi a0,a0,-258
     ret

This patch solves the mentioned issue.

RISC-V: avoid splitting small constant in *bclri<mode>_nottwobits
        and *bclridisi_nottwobit patterns

gcc/
        * config/riscv/bitmanip.md Updated predicats of
        bclri<mode>_nottwobits and bclridisi_nottwobits patterns
        * config/riscv/predicates.md (not_uimm_extra_bit_or_nottwobits):
        Adjust predicate to avoid splitting arith constants
        * config/riscv/predicates.md (const_nottwobits_not_arith_operand):
        New predicate

gcc/testsuite
        * gcc.target/riscv/zbs-bclri-nottwobits.c: New test.