Pass correct type to irange::contains_p() in ipa-cp.cc.
There is a call to contains_p() in ipa-cp.cc which passes incompatible
types.  This currently works because deep in the call chain, the legacy
code uses tree_int_cst_lt which performs the operation with
widest_int.  With the upcoming removal of legacy, contains_p() will be
stricter.

OK pending tests?

gcc/ChangeLog:

	* ipa-cp.cc (ipa_range_contains_p): New.
	(decide_whether_version_node): Use it.