tree-optimization/109609 - correctly interpret arg size in fnspec
By majority vote and a hint from the API name which is
arg_max_access_size_given_by_arg_p this interprets a memory access
size specified as given as other argument such as for strncpy
in the testcase which has "1cO313" as specifying the _maximum_
size read/written rather than the exact size.  There are two
uses interpreting it that way already and one differing.  The
following adjusts the differing and clarifies the documentation.

Bootstrap & regtest running on x86_64-unknown-linux-gnu.

It looks like arg_access_size_given_by_type_p is interpreted in
a similar way but the API doesn't reflect this (no _max).  On IRC
we discussed it would be nice to have both, exact and non-exact
size (for example use 'a'..'j' for the alternate variant).

I still wonder if my interpretation is correct in that '1'..'9'
specify a _bound_ for the size.  Honza, do you remember why
you wrote the IPA modref users this way?  Was that intentional?
Thus is my reverse engineering of the desired semantics correct?

Thanks,
Richard.

	PR tree-optimization/109609
	* attr-fnspec.h (arg_max_access_size_given_by_arg_p):
	Clarify semantics.
	* tree-ssa-alias.cc (check_fnspec): Correctly interpret
	the size given by arg_max_access_size_given_by_arg_p as
	maximum, not exact, size.

	* gcc.dg/torture/pr109609.c: New testcase.