Unloop no longer looping loops in loop-ch
Hi,
I noticed this after adding sanity check that the upper bound on number
of iterations never drop to -1.  It seems to be relatively common case
(happening few hundred times in testsuite and also during bootstrap)
that loop-ch duplicates enough so the loop itself no longer loops.

This is later detected in loop unrolling but since we test the number
of iterations anyway it seems better to do that earlier.

Bootstrapped/regtested x86_64-linux, OK?
Honza

	* cfgloopmanip.h (unloop_loops): Declare.
	* tree-ssa-loop-ch.cc (ch_base::copy_headers): Unloop loops that
	no longer loop.
	* tree-ssa-loop-ivcanon.cc (unloop_loops): Export; do not release
	vectors passed.
	(canonicalize_induction_variables): Release vectors here.
	(tree_unroll_loops_completely): Release vectors here.