avr: Set param_min_pagesize to 0 [PR105523]
Hi,

This patch fixes PR 105523 by setting param_min_pagesize to 0 for the
avr target. For this target, zero and offsets from zero are perfectly
valid addresses, and the default value of param_min_pagesize ends up
triggering warnings on valid memory accesses.

Ok for trunk and backporting to 13 and 12 branches?

Regards
Senthil

	PR target/105523

	gcc/ChangeLog:

			* config/avr/avr.cc (avr_option_override): Set
			param_min_pagesize to 0.

	gcc/testsuite/ChangeLog:

			* gcc.target/avr/pr105523.c: New test.